+++ src/otx/algo/detection/utils/utils.py
@@ -177,3 +177,9 @@ class SigmoidGeometricMean(Function):
 
 
 sigmoid_geometric_mean = SigmoidGeometricMean.apply
+
+
+def inverse_sigmoid(x: torch.Tensor, eps: float = 1e-5) -> torch.Tensor:
+    """Compute the inverse of sigmoid function."""
+    x = x.clip(min=0.0, max=1.0)
+    return torch.log(x.clip(min=eps) / (1 - x).clip(min=eps))