+++ src/otx/core/utils/utils.py
@@ -5,9 +5,10 @@
 
 from __future__ import annotations
 
+import importlib
 from collections import defaultdict
 from multiprocessing import cpu_count
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 import torch
 from datumaro.components.annotation import AnnotationType, LabelCategories
@@ -84,3 +85,19 @@ def get_idx_list_per_classes(dm_dataset: DmDataset, use_string_label: bool = Fal
     for k in stats:
         stats[k] = list(dict.fromkeys(stats[k]))
     return stats
+
+
+def import_object_from_module(obj_path: str) -> Any:  # noqa: ANN401
+    """Get object from import format string."""
+    module_name, obj_name = obj_path.rsplit(".", 1)
+    module = importlib.import_module(module_name)
+    return getattr(module, obj_name)
+
+
+def remove_state_dict_prefix(state_dict: dict[str, Any], prefix: str) -> dict[str, Any]:
+    """Remove prefix from state_dict keys."""
+    new_state_dict = {}
+    for key, value in state_dict.items():
+        new_key = key.replace(prefix, "")
+        new_state_dict[new_key] = value
+    return new_state_dict