+++ CHANGELOG.md
@@ -6,9 +6,34 @@ All notable changes to this project will be documented in this file.
 
 ### New features
 
+- Add RT-DETR model for object detection task
+  (https://github.com/openvinotoolkit/training_extensions/pull/3741)
+- Add Multi-Label & H-label Classification with torchvision models
+  (https://github.com/openvinotoolkit/training_extensions/pull/3697)
+- Add Hugging-Face Model Wrapper for Classification
+  (https://github.com/openvinotoolkit/training_extensions/pull/3710)
+- Add Hugging-Face Model Wrapper for Detection
+  (https://github.com/openvinotoolkit/training_extensions/pull/3747)
+- Add Hugging-Face Model Wrapper for Semantic Segmentation
+  (https://github.com/openvinotoolkit/training_extensions/pull/3749)
+- Enable torch.compile to work with classification
+  (https://github.com/openvinotoolkit/training_extensions/pull/3758)
+
 ### Enhancements
 
+- Enable to override data configurations
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3748>)
+- Enable to use input_size at transforms in recipe
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3759>)
+- Enable to use polygon and bitmap mask as prompt inputs for zero-shot learning
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3769>)
+
 ### Bug fixes
+
+- Fix Combined Dataloader & unlabeled warmup loss in Semi-SL
+  (https://github.com/openvinotoolkit/training_extensions/pull/3723)
+- Revert #3579 to fix issues with replacing coco_instance with a different format in some dataset
+  (https://github.com/openvinotoolkit/training_extensions/pull/3753)
 
 ## \[v2.1.0\]